+++ cwms-radar-client/build.gradle
@@ -1,71 +1,15 @@
-plugins {
-    id 'org.hidetake.swagger.generator' version '2.18.2'
-//    id "com.github.eirnym.js2p" version "1.0"
-}
-
 group = "mil.army.usace.hec"
 version = "1.0-SNAPSHOT"
 
 
 dependencies {
     implementation(project(":cwms-http-client"))
+    implementation(project(":cwms-radar-model"))
     implementation(platform("com.squareup.okhttp3:okhttp-bom:4.9.2"))
     implementation("com.squareup.okhttp3:okhttp")
-    implementation("com.fasterxml.jackson.core:jackson-databind:2.13.0")
-    implementation("com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.13.0")
-    implementation("io.swagger:swagger-annotations:1.6.3")
-    implementation("javax.validation:validation-api:2.0.1.Final")
 
     testImplementation("com.squareup.okhttp3:mockwebserver")
     testImplementation("org.junit.jupiter:junit-jupiter-api:5.8.1")
     testImplementation("com.squareup.okhttp3:mockwebserver:4.9.0")
     testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:5.8.1")
-
-    swaggerCodegen("io.swagger.codegen.v3:swagger-codegen-cli:3.0.5")
-//    compileOnly('org.jsonschema2pojo:jsonschema2pojo-gradle-plugin:1.1.1')
-//    compileOnly('io.swagger.codegen.v3:swagger-codegen-maven-plugin:3.0.16')
-}
-
-//buildscript {
-//    repositories {
-//        mavenCentral()
-//    }
-//
-//    dependencies {
-//        classpath('org.jsonschema2pojo:jsonschema2pojo-gradle-plugin:1.1.1')
-//        classpath('io.swagger.codegen.v3:swagger-codegen-maven-plugin:3.0.16')
-//    }
-//
-//}
-
-//jsonSchema2Pojo {
-//    targetPackage = 'mil.army.usace.hec.cwms.radar.client'
-//    source = files("${sourceSets.main.output.resourcesDir}/radar/json")
-//    targetDirectory = file("${project.buildDir}/generated-sources/js2p")
-//    sourceType = 'json'
-//    propertyWordDelimiters = '-_'
-//    targetVersion = "1.8"
-//
-//    annotationStyle = 'jackson'
-//    includeHashcodeAndEquals = false
-//    includeToString = false
-//    initializeCollections = true
-//    includeGetters = true
-//    includeSetters = false
-//    dateType = "java.time.LocalDate"
-//    dateTimeType = "java.time.LocalDateTime"
-//    timeType = "java.time.LocalTime"
-//    includeAdditionalProperties = false
-//
-//}
-
-swaggerSources {
-    cwmsRadar {
-        inputFile = file("/Users/adamkorynta/Documents/RMA/Git/adamkorynta/cwms-radar-client/cwms-radar-client/src/test/resources/radar/json/cwms-radar-swagger.yaml")
-        code {
-            language = 'spring'
-//            components = ['models']
-//            configFile = file("/Users/adamkorynta/Documents/RMA/Git/adamkorynta/cwms-radar-client/cwms-radar-client/src/test/resources/radar/json/cwms-radar-swagger.yaml")
-        }
-    }
 }
