--- cwms-radar-client/src/main/resources/radar/json/locations.json
@@ -1,58 +1,0 @@
-{
-  "locations": {
-    "query-info": {
-      "processed-at": "oralind19c:V19CSWT181LRTS",
-      "time-of-query": "2021-10-28T21:52:17Z",
-      "process-query": "PT2.372S",
-      "format-output": "PT0.663S",
-      "requested-format": "JSON",
-      "requested-office": "SWT",
-      "requested-items": [
-        {
-          "name": "*",
-          "unit": "SI",
-          "datum": "NAVD88"
-        }
-      ],
-      "total-locations-retrieved": 6711,
-      "unique-locations-retrieved": 3361
-    },
-    "locations": [
-      {
-        "identity": {
-          "office": "SWT",
-          "name": "000512",
-          "alternate-names": []
-        },
-        "label": {
-          "public-name": "Resource Management Associates",
-          "long-name": "Resource Management Associates - Water Resources Engineering",
-          "description": "At RMA, our team of engineers and software developers create and apply advanced numerical models and software systems that support water resource management and environmental stewardship."
-        },
-        "geolocation": {
-          "latitude": 38.56326,
-          "longitude": -121.7303,
-          "horizontal-datum": "NAD83",
-          "elevation": {
-            "value": 42.07645,
-            "unit": "m",
-            "datum": "NAVD88",
-            "estimate": "true"
-          }
-        },
-        "political": {
-          "nation": "UNITED STATES",
-          "state": "CA",
-          "county": "Yolo",
-          "timezone": "America/Los_Angeles",
-          "nearest-city": "Davis",
-          "bounding-office": "SPD"
-        },
-        "classification": {
-          "location-kind": "SITE",
-          "location-type": "Contractor"
-        }
-      }
-    ]
-  }
-}
