+++ cwms-radar-model/src/main/java/mil/army/usace/hec/cwms/radar/client/model/RadarObjectMapper.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2021
+ * United States Army Corps of Engineers - Hydrologic Engineering Center (USACE/HEC)
+ * All Rights Reserved.  USACE PROPRIETARY/CONFIDENTIAL.
+ * Source may not be released without written approval from HEC
+ */
+
+package mil.army.usace.hec.cwms.radar.client.model;
+
+import com.fasterxml.jackson.core.json.JsonReadFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import java.io.IOException;
+
+/**
+ *
+ */
+public final class RadarObjectMapper {
+
+    private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper()
+        .registerModule(new JavaTimeModule())
+        .configure(JsonReadFeature.ALLOW_UNESCAPED_CONTROL_CHARS.mappedFeature(), true);
+
+    public static <T> T mapJsonToObject(String json, Class<T> classObject) throws IOException {
+        return OBJECT_MAPPER.readValue(json, classObject);
+    }
+}
