+++ cwms-radar-model/src/main/java/mil/army/usace/hec/cwms/radar/client/model/TimeSeries.java
