+++ settings.gradle
@@ -2,3 +2,4 @@ rootProject.name = 'cwms-radar-client-parent'
 
 include ":cwms-http-client"
 include ":cwms-radar-client"
+include ":cwms-radar-model"
